Add PATCH methods

+++ src/Structures/ResourceIdentifier.php
@@ -31,6 +31,9 @@ class ResourceIdentifier
      */
     protected $relationships;
 
+    /**
+     * @var string[] Array to keep track of relationships that are changed.
+     */
     protected $changedRelationships = [];
 
     /**
@@ -53,7 +56,7 @@ class ResourceIdentifier
      *
      * @return string The resource type.
      */
-    public function type()
+    public function type() : string
     {
         return $this->resourceType;
     }
@@ -63,7 +66,7 @@ class ResourceIdentifier
      *
      * @return string|null The resource Id.
      */
-    public function id()
+    public function id() : ?string
     {
         return $this->id;
     }
@@ -72,8 +75,6 @@ class ResourceIdentifier
      * Make a GET request to the resource.
      *
      * @return Resource|ResourceSet A resource or resource set.
-     * @throws ExonetApiException The exception.
-     *
      */
     public function get()
     {
@@ -83,9 +84,28 @@ class ResourceIdentifier
     /**
      * Delete this resource from the API.
      */
-    public function delete()
+    public function delete() : void
     {
-        $this->request->delete($this->id());
+        // If there are no changed relationships, perform a 'normal' delete.
+        if (empty($this->changedRelationships)) {
+            $this->request->delete($this->id());
+
+            return;
+        }
+
+        // If there are changed relationships, transform them to JSON and send a DELETE to the relationship endpoint.
+        foreach ($this->changedRelationships as $relationship) {
+            $relationData = $this->relationship($relationship);
+            if (is_array($relationData)) {
+                array_walk($relationData, function (&$relationItem) {
+                    $relationItem = $relationItem->toJson();
+                });
+            } else {
+                $relationData = $relationData->toJson();
+            }
+
+            $this->request->delete($this->id().'/relationships/'.$relationship, ['data' => $relationData]);
+        }
     }
 
     /**
@@ -103,33 +123,45 @@ class ResourceIdentifier
     /**
      * Get a relationship definition to another resource.
      *
-     * @param string                     $name The name of the relationship.
-     * @param ApiResourceIdentifier|null $resource
+     * @param string                                  $name The name of the relationship.
+     * @param ResourceIdentifier|ResourceIdentifier[] $resource
      *
-     * @return Relation|Relationship|$this
+     * @return Relation|Relationship|$this The requested relation data or the current resource when setting a relation.
      */
     public function relationship(string $name, $resource = null)
     {
-        // Check if the relationship is already defined. If not, create it now.
-        if (!isset($this->relationships[$name])) {
-            $this->relationships[$name] = new Relationship($name, $this->type(), $this->id());
+        // If there are is only a single argument, get the relation.
+        if (func_num_args() === 1) {
+            // Check if the relationship is already defined. If not, create it now.
+            if (!isset($this->relationships[$name])) {
+                $this->relationships[$name] = new Relationship($name, $this->type(), $this->id());
+            }
+
+            return $this->relationships[$name];
         }
 
-        // If there are two arguments given, set the value.
-        if (func_num_args() === 2) {
-            $this->relationships[$name]->setResourceIdentifiers($resource);
-            $this->changedRelationships[] = $name;
-
-            return $this;
+        // Set the relation data.
+        if (is_array($resource)) {
+            foreach ($resource as $resourceIdentifier) {
+                $this->relationships[$name][] = $resourceIdentifier;
+            }
+        } else {
+            $this->relationships[$name] = $resource;
         }
 
-        return $this->relationships[$name];
-    }
-
-    public function resetChangedRelationhips()
-    {
-        $this->changedRelationships = [];
+        $this->changedRelationships[] = $name;
 
         return $this;
+    }
+
+    /**
+     * Transform the set identifiers for this resource to an array that can be used for JSON.
+     */
+    protected function toJson() : array
+    {
+        return [
+            'type' => $this->type(),
+            'id' => $this->id(),
+        ];
     }
 }
